requirements.txt
Django==5.2.11
django-cors-headers==4.9.0
Pillow==11.0.0
numpy==1.26.4
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
